+++ src/io/flutter/utils/JxBrowserUtils.java
@@ -25,18 +25,22 @@ public class JxBrowserUtils {
     String name = "";
     final boolean is64Bit = Objects.requireNonNull(CpuArch.CURRENT).width == 64;
     if (SystemInfo.isMac) {
-      if (SystemInfo.isAarch64){
+      if (CpuArch.isArm64()) {
         name = "mac-arm";
-      } else {
+      }
+      else {
         name = "mac";
       }
-    } else if (SystemInfo.isWindows) {
+    }
+    else if (SystemInfo.isWindows) {
       if (CpuArch.is32Bit()) {
         name = "win32";
-      } else if (is64Bit) {
+      }
+      else if (is64Bit) {
         name = "win64";
       }
-    } else if (SystemInfo.isLinux && is64Bit) {
+    }
+    else if (SystemInfo.isLinux && is64Bit) {
       name = "linux64";
     }