+++ testSrc/unit/io/flutter/sdk/FlutterSdkVersionTest.java
@@ -18,43 +18,18 @@ public class FlutterSdkVersionTest {
 
   @Test
   public void trackSdkVersionSupport() {
-    assertFalse(new FlutterSdkVersion("3.9.0").isSDKSupported());
-    assertFalse(new FlutterSdkVersion("3.9.0").isSDKSupported());
-    assertFalse(new FlutterSdkVersion("3.9.0.pre").isSDKSupported());
-    assertFalse(new FlutterSdkVersion( "3.10.0.pre").isSDKSupported());
-    assertFalse(new FlutterSdkVersion( "3.10.1").isSDKSupported());
-    assertFalse(new FlutterSdkVersion( "3.10.2").isSDKSupported());
-    assertFalse(new FlutterSdkVersion( "3.10.3").isSDKSupported());
-    assertFalse(new FlutterSdkVersion( "3.10.4").isSDKSupported());
-    assertFalse(new FlutterSdkVersion("3.10.3").isSDKSupported());
-    assertFalse(new FlutterSdkVersion("3.10.4").isSDKSupported());
-    assertFalse(new FlutterSdkVersion("3.10.5").isSDKSupported());
-    assertFalse(new FlutterSdkVersion("3.10.6").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.0").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.1").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.2").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.3").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.4").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.5").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.6").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.7").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.8").isSDKSupported());
-    assertTrue(new FlutterSdkVersion("3.13.9").isSDKSupported());
-    assertFalse(new FlutterSdkVersion( "unknown").isSDKSupported());
+    assertFalse(new FlutterSdkVersion("3.15.0").isSDKSupported());
+    assertFalse(new FlutterSdkVersion("3.16.0-0.1.pre").isSDKSupported());
+    assertTrue(new FlutterSdkVersion("3.16.0").isSDKSupported());
+    assertTrue(new FlutterSdkVersion("3.16.1").isSDKSupported());
+    assertTrue(new FlutterSdkVersion("3.17.0").isSDKSupported());
+    assertFalse(new FlutterSdkVersion("unknown").isSDKSupported());
 
-    assertTrue(new FlutterSdkVersion("3.13.0").isSDKAboutToSunset());
-    assertTrue(new FlutterSdkVersion("3.13.1").isSDKAboutToSunset());
-    assertTrue(new FlutterSdkVersion("3.13.2").isSDKAboutToSunset());
-    assertTrue(new FlutterSdkVersion("3.13.3").isSDKAboutToSunset());
-    assertTrue(new FlutterSdkVersion("3.13.4").isSDKAboutToSunset());
-    assertTrue(new FlutterSdkVersion("3.13.5").isSDKAboutToSunset());
-    assertTrue(new FlutterSdkVersion("3.13.6").isSDKAboutToSunset());
-    assertTrue(new FlutterSdkVersion("3.13.7").isSDKAboutToSunset());
-    assertTrue(new FlutterSdkVersion("3.13.8").isSDKAboutToSunset());
-    assertTrue(new FlutterSdkVersion("3.13.9").isSDKAboutToSunset());
-    assertFalse(new FlutterSdkVersion("3.16.0").isSDKAboutToSunset());
-    assertFalse(new FlutterSdkVersion("3.16.1").isSDKAboutToSunset());
-    assertFalse(new FlutterSdkVersion("3.16.2").isSDKAboutToSunset());
+    assertTrue(new FlutterSdkVersion("3.16.0").isSDKAboutToSunset());
+    assertTrue(new FlutterSdkVersion("3.17.0").isSDKAboutToSunset());
+    assertTrue(new FlutterSdkVersion("3.19.3").isSDKAboutToSunset());
+    assertFalse(new FlutterSdkVersion("3.19.4").isSDKAboutToSunset());
+    assertFalse(new FlutterSdkVersion("3.20.0").isSDKAboutToSunset());
   }
 
   @Test